requirements.txt
python-dotenv
yfinance
pandas
matplotlib
requests
textblob
streamlit